Tolerate v1 certs without basic constraints (#53)
Our current basic constraints checking policy requires the presence of
the basic constraints extension to be a CA. That's too aggressive in the
face of the ongoing persistence of v1 certs, both in the root store and
in the wider ecosystem of devices. We can safely tolerate v1
certificates in almost all cases.