X509: make the code more portable
`CoreFoundation` is not available on all platforms.  It never was meant
to be available on Linux, but was accidentally pushed into the release
and is now stuck.  Windows intentionally did not make this module
visible.  With the future work with the Foundation rewrite
`CoreFoundation` will disappear in more locations.

Windows does not support `inet_pton` except unless the string is
explicitly ANSI and not UTF-8.  We must use the proper spelling of
`InetPtonW` which takes a UTF-16 string.

The import of `CoreFoundation` was more for the dependency on the C
library as a side-effect.  On Windows, the networking functions are
provided by a separate user space library rather than the C library.
This is now highlighted in the imports.