CMake: Use full module triples (#252)
Teach the CMake build to install swiftmodules using the full module
triple. This uses the compiler to determine the appropriate module
triple instead of attempting to compute the architecture in CMake
script, making the build more robust for new platforms and
architectures.

Also removing redundant policies. CMP0077 and CMP0091 were added before
the minimum CMake version and therefore will always be active in this
project.