Support singing via Certificate PrivateKey (#257)
This PR introduces two key changes to improve flexibility when working
with signatures in swift-certificates. Previously, it was only possible
to verify signatures using `Certificate.PrivateKey`, but not to generate
them.

- The internal `sign(bytes:signatureAlgorithm:)` method on
`Certificate.PrivateKey` is now public, allowing users to generate
signatures directly.
- `Certificate.Signature` gains a new `rawRepresentation` property that
exposes the raw byte representation of a signature.