Require VerifierPolicy not have isolated conformance
Motivation

While conceptually it is possible for verifier policies to
have isolated conformance, in practice the way PolicyBuilder is
implemented makes that nearly impossible. Given that the utility
of an isolated VerifierPolicy is very limited, let's just
disallow it.

Modifications

Require VerifierPolicy to have a SendableMetatype

Result

Easier to build correct compositions over PolicyBuilder