Expose supported signature algorithms for privkeys
Motivation:

When loading a private key from disk as a Certificate.PrivateKey there
is no straight forward way to query the supported signature
algorithms, e.g., to perform a signature with the key. For signature
algorithm negotiation, the raw values defined in RFC can also be useful.

Modifications:

Add a new property to query the supported signature algorithms for a
given key. Add a new property that translates the signature algorithms
to the value defined in RFC 8446.

Results:

Working with certificate private keys and signatures gets easier.