Expose supported signature algorithms for private keys (#281)
### Motivation:

When loading a private key from disk as a `Certificate.PrivateKey` there
is no straight forward way to query the supported signature algorithms.
For signature algorithm negotiation, the raw values defined in RFC are
useful as well.

### Modifications:

Add a new property to query the supported signature algorithms for a
given key. Add a new property that translates a signature algorithm to
the respective value defined in RFC 8446.

### Results:

Working with certificate private keys and signature algorithms gets
easier.