Add validation function for raw signatures (#284)
### Motivation

Signatures can be performed on any kind of data, e.g., to prove
ownership of the private key for a given certificate. This data usually
comes as a blob of bytes.

### Modification

Add a signature verification function that operates on bytes represented
by the `DataProtocol`. The bytes are converted to the signature type
matching the key and share an implementation with the verification
function for `Certificate.Signature`.

### Result

Easy validation of signatures represented as bytes.

---------

Co-authored-by: Cory Benfield <lukasa@apple.com>